Fix premature loading of AutoInstall
* This bug prevented running configuration clients during 2nd stage
* Fixes bsc#944770
