+++ package/autoyast2.changes
@@ -1,4 +1,11 @@
 -------------------------------------------------------------------
+Mon Sep  7 22:32:50 UTC 2015 - igonzalezsosa@suse.com
+
+- Fix premature loading of AutoInstall which prevented running
+  configuration clients during 2nd stage (bsc#944770)
+- 3.1.93
+
+-------------------------------------------------------------------
 Tue Sep  1 15:25:44 UTC 2015 - igonzalezsosa@suse.com
 
 - Move users creation to the first stage, so it is not needed
