+++ src/modules/AutoInstall.rb
@@ -20,8 +20,6 @@ module Yast
       Yast.import "AutoInstallRules"
       Yast.import "Report"
       Yast.import "TFTP"
-      Yast.import "ProductControl"
-      Yast.import "Mode"
 
       @autoconf = false
       AutoInstall()
@@ -303,27 +301,6 @@ module Yast
       true
     end
 
-    # Determines if the second stage should be executed
-    #
-    # Checks Mode, AutoinstConfig and ProductControl to decide if it's
-    # needed.
-    #
-    # FIXME: It's almost equal to InstFunctions.second_stage_required?
-    # defined in yast2-installation, but exists to avoid a circular dependency
-    # between packages (yast2-installation -> autoyast2-installation).
-    #
-    # @return [Boolean] 'true' if it's needed; 'false' otherwise.
-    def second_stage_required?
-      return false unless Stage.initial
-      if (Mode.autoinst || Mode.autoupgrade) && !AutoinstConfig.second_stage
-        Builtins.y2milestone("Autoyast: second stage is disabled")
-        false
-      else
-        ProductControl.RunRequired("continue", Mode.mode)
-      end
-    end
-    alias_method :second_stage_required, :second_stage_required?
-
     publish :variable => :autoconf, :type => "boolean"
     publish :function => :callbackTrue_boolean_string, :type => "boolean (string)"
     publish :function => :callbackFalse_boolean_string, :type => "boolean (string)"
@@ -347,7 +324,6 @@ module Yast
     publish :function => :Save, :type => "boolean ()"
     publish :function => :Finish, :type => "void (string)"
     publish :function => :PXELocalBoot, :type => "boolean ()"
-    publish function: :second_stage_required, type: "boolean ()"
   end
 
   AutoInstall = AutoInstallClass.new
